requirements.txt
paramiko>=3.4.0
google-auth>=2.29.0
google-api-python-client>=2.126.0
pandas>=2.2.0
streamlit>=1.33.0
